requirements.txt
Django==1.5.1
django-annoying==0.7.7
South==0.8.1
coverage==3.6
